+++ lib/ua_inspector/util.ex
@@ -5,13 +5,24 @@ defmodule UAInspector.Util do
   Generate a regex to be used for engine version detection.
   """
   @spec build_engine_regex(name :: String.t()) :: Regex.t()
+  def build_engine_regex("Clecko") do
+    # sigil_S used to ensure escaping is kept as-is
+    # Concatenated expression:
+    # - [ ](?:rv[: ]([0-9\.]+)).*(?:g|cl)ecko\/[0-9]{8,10}
+    # - Regular expression of `build_engine_regex("Clecko")`
+    Regex.compile!(
+      ~S"(?:[ ](?:rv[: ]([0-9\.]+)).*(?:g|cl)ecko\/[0-9]{8,10}|Gecko\s*\/?\s*((?(?=\d+\.\d)\d+[.\d]*|\d{1,7}(?=(?:\D|$)))))",
+      [:caseless]
+    )
+  end
+
   def build_engine_regex("Gecko") do
     # sigil_S used to ensure escaping is kept as-is
     # Concatenated expression:
-    # - [ ](?:rv[: ]([0-9\.]+)).*Gecko\/[0-9]{8,10}
+    # - [ ](?:rv[: ]([0-9\.]+)).*(?:g|cl)ecko\/[0-9]{8,10}
     # - Regular expression of `build_engine_regex("Gecko")`
     Regex.compile!(
-      ~S"(?:[ ](?:rv[: ]([0-9\.]+)).*Gecko\/[0-9]{8,10}|Gecko\s*\/?\s*((?(?=\d+\.\d)\d+[.\d]*|\d{1,7}(?=(?:\D|$)))))",
+      ~S"(?:[ ](?:rv[: ]([0-9\.]+)).*(?:g|cl)ecko\/[0-9]{8,10}|Gecko\s*\/?\s*((?(?=\d+\.\d)\d+[.\d]*|\d{1,7}(?=(?:\D|$)))))",
       [:caseless]
     )
   end