+++ lib/ua_inspector/util.ex
@@ -27,6 +27,7 @@ defmodule UAInspector.Util do
     )
   end
 
+  def build_engine_regex("Arachne"), do: build_engine_regex("Arachne\\/5\\.")
   def build_engine_regex("Blink"), do: build_engine_regex("Chrome|Cronet")
 
   def build_engine_regex(name) do