+++ lib/ua_inspector/parser/os.ex
@@ -20,13 +20,13 @@ defmodule UAInspector.Parser.OS do
   ]
 
   @platforms [
-    {"ARM", Util.build_regex("arm|aarch64|Apple ?TV|Watch ?OS|Watch1,[12]")},
+    {"ARM", Util.build_regex("arm|.*arm64|aarch64|Apple ?TV|Watch ?OS|Watch1,[12]")},
     {"SuperH", Util.build_regex("sh4")},
     {"SPARC64", Util.build_regex("sparc64")},
     {"LoongArch64", Util.build_regex("loongarch64")},
     {"MIPS", Util.build_regex("mips")},
     {"x64", Util.build_regex("64-?bit|WOW64|(?:Intel)?x64|WINDOWS_64|win64|amd64|x86_?64")},
-    {"x86", Util.build_regex(".+32bit|.+win32|(?:i[0-9]|x)86|i86pc")}
+    {"x86", Util.build_regex(".*32bit|.*win32|(?:i[0-9]|x)86|i86pc")}
   ]
 
   @impl UAInspector.Parser.Behaviour