+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - Upstream improvements for "360 Secure" browser detection have been integrated
     - Upstream improvements for "PICO OS" detection have been integrated
     - Upstream improvements for "SPARC64" platform detection have been integrated
+    - Upstream improvements for wearable device detection have been integrated
 
 ## v3.9.0 (2024-02-20)