+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enhancements
     - Upstream improvements for "360 Secure" browser detection have been integrated
     - Upstream improvements for "PICO OS" detection have been integrated
+    - Upstream improvements for "QJY TV" browser detection have been integrated
     - Upstream improvements for "SPARC64" platform detection have been integrated
     - Upstream improvements for wearable device detection have been integrated