+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Upstream improvements for "Arachne" browser engine detection have been integrated
     - Upstream improvements for "Clecko" browser engine detection have been integrated
     - Upstream improvements for "Edge WebView" detection have been integrated
+    - Upstream improvements for "LibWeb" browser engine detection have been integrated
     - Upstream improvements for "Lineage OS" detection have been integrated
     - Upstream improvements for "PICO OS" detection have been integrated
     - Upstream improvements for "QJY TV" browser detection have been integrated