+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     - Upstream improvements for "Edge WebView" detection have been integrated
     - Upstream improvements for "LibWeb" browser engine detection have been integrated
     - Upstream improvements for "Lineage OS" detection have been integrated
+    - Upstream improvements for "LoongArch64" platform detection have been integrated
     - Upstream improvements for "PICO OS" detection have been integrated
     - Upstream improvements for "QJY TV" browser detection have been integrated
     - Upstream improvements for "TV Bro" browser detection have been integrated