+++ CHANGELOG.md
@@ -13,12 +13,12 @@
     - Upstream improvements for "Edge WebView" detection have been integrated
     - Upstream improvements for "LibWeb" browser engine detection have been integrated
     - Upstream improvements for "Lineage OS" detection have been integrated
-    - Upstream improvements for "LoongArch64" platform detection have been integrated
     - Upstream improvements for "PICO OS" detection have been integrated
     - Upstream improvements for "QJY TV" browser detection have been integrated
     - Upstream improvements for "TV Bro" browser detection have been integrated
     - Upstream improvements for "SPARC64" platform detection have been integrated
     - Upstream improvements for Android mobile/tablet device detection have been integrated
+    - Upstream improvements for OS platform detection have been integrated
     - Upstream improvements for wearable device detection have been integrated
 
 ## v3.9.0 (2024-02-20)